% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answer_key_test_statistic_value.R
\name{answer_key_test_statistic_value}
\alias{answer_key_test_statistic_value}
\title{Test Statistic Value of Edge Avoidance Hypothesis Test (Answer Key)}
\usage{
answer_key_test_statistic_value(exam, numchoicesperitem = NULL,
  check_format = TRUE)
}
\arguments{
\item{exam}{A data.frame object containing a test that follows the formatting of a typical Scantron file.}

\item{numchoicesperitem}{A vector which has the same length as the number of questions on the test, where each element in
the vector corresponds to the number of choices there were for each of the questions on the test. For example, if your
test had 35 questions which each had 5 options, numchoicesperitem = rep(5, 35).}

\item{check_format}{a logical value (default = TRUE) indicating whether the "exam" object should be tested
for correct formatting.}
}
\value{
Returns a numeric value which corresponds to the test statistic value for the hypothesis test.
}
\description{
Finds the test statistic value of a simple hypothesis test, testing whether there is significant
statistical evidence of the answer key avoiding edge options.
}
\details{
The null hypothesis is that the answer key does not avoid edge options. That is, the true proportion of
edge options in the answer key is the expected proportion of edge options in the answer key. For example, if every question
on the test had 5 options, the expected proportion of edge options in the answer key would be 0.4. The alternative hypothesis
is that the answer key does avoid edge options (that is, the true proportion of edge options in the answer key is less than
the expected proportion of edge options in the answer key.)
}
\note{
The argument "exam" must be in the format of a typical Scantron results file.
Column 1 should correspond to ID (e.g., student number); column  2 should correspond to
DEPT (e.g., MATH); column 3 should correspond to COURSE CODE (e.g., 1051); the remaining
columns should each correspond to one of the questions on the test. The header of the data
frame should contain the column names, and row 1 of the data frame should contain the answer
key for the test. For example, if you had an exam with 25 students and 40 questions,
the data.frame object should have 26 rows and 43 columns.

The "check_format" argument defaults to null. If this is left as null, the function will call
the \code{\link{num_choices_per_item}} function, which will do its best to guess the number of options
for each question. This is done by looking at the student answers and finding the
"largest" answer for each question. For example, if at least one student answered "E", but no students
answered "F", the function would guess that there were 5 options for that question.
}
