% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/letter2num.R
\name{letter2num}
\alias{letter2num}
\title{Convert a Letter to a Number}
\usage{
letter2num(x)
}
\arguments{
\item{x}{An uppercase letter.}
}
\value{
Returns a numeric corresponding to the position in the alphabet of the letter given.
}
\description{
Finds the numeric position in the alphabet of a letter. For example, "A" = 1, "B" = 2, and so on.
}
